Added animations for eye candy
Added animations for some eye candy. There is a bug where the score wont show/hide, so I hid it for now.

There is some cleanup needed such as the tempContainer and the overall container attributes since we delete the text after checking and re-add it again. This won't be in an RC for now.